requirements.txt
python-telegram-bot==13.7
requests==2.30.0
googletrans==3.1.0a0
gTTS==2.3.2
python-dotenv==0.19.0
